Migrate hex funcs
