+++ src/common/common.ts
@@ -1,5 +1,3 @@
-import { toHexString } from './hex-strings'
-
 export const assert = (condition: () => boolean, reason?: string) => {
   try {
     if (condition() === false) {
@@ -7,21 +5,5 @@ export const assert = (condition: () => boolean, reason?: string) => {
     }
   } catch (err) {
     throw new Error(`Assertion failed: ${reason}\n${err}`)
-  }
-}
-
-export const toRpcHexString = (n: number): string => {
-  if (n === 0) {
-    return '0x0'
-  } else {
-    return '0x' + toHexString(n).slice(2).replace(/^0+/, '')
-  }
-}
-
-export const padHexString = (str: string, length: number): string => {
-  if (str.length === 2 + length * 2) {
-    return str
-  } else {
-    return '0x' + str.slice(2).padStart(length * 2, '0')
   }
 }
